Add files via upload
You type the words "!show" before you type in the image you want. 
For example, you do: !show duck and it will show you an image of a duck.
The word "test" doesn't work in this code, as it creates and shows you an error.
Thank you to whoever downloaded this code, and good luck.